% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication.R
\name{getPublicationMetadata}
\alias{getPublicationMetadata}
\title{Download metadata from a publication from a DOI}
\usage{
getPublicationMetadata(dois)
}
\description{
Uses the R package rcrossref to query the CrossRef API.
}
