% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{screenshot}
\alias{screenshot}
\title{Save a thumbnail screenshot of a website}
\usage{
screenshot(websites, thumbnails)
}
\arguments{
\item{websites}{character vector of URLs to websites}

\item{thumbnails}{character vector of filenames for thumbnail images}
}
\value{
Character vector of filenames of thumbnail images (invisibly)
}
\description{
Save a thumbnail screenshot of a website
}
\examples{
  screenshot("https://stephenslab.github.io/wflow-divvy/", "test.png")

}
